In-progress work on Coz

+++ benchmarks/word_count/sort-pthread.c
@@ -102,7 +102,7 @@ void sort_pthreads(void *base, size_t num_elems, size_t width,
 	   /* Barrier, wait for all threads to finish */
 	   for (i = 0; i < num_threads; i++)
 	   {
-		  int ret_val;
+		  long ret_val;
 		  CHECK_ERROR(pthread_join(tid[i], (void **)(void*)&ret_val) != 0);
 		  CHECK_ERROR(ret_val != 0);
 	   }
